Remove unnecessary parallel data structure
Previously, I thought this would be necessary to unwind the visited vertices,
but considering that a synset can have multiple hypernyms, there is no guarantee
that traversing backwards through `edgeTo` would have the desired effect of
unmarking all previously-marked vertices.

Accordingly, just use a stack and push onto that stack the ID of all the vertices
which were modified.