Use a set to track added nouns
Previously, no checks were made to see whether a noun had already been
added to the `allNouns` list.

Unfortunately, there are multiple instances of single nouns.

Accordingly, use a TreeSet to check whether the current noun has already
been added to the `allNouns` list.

This has the added benefit of providing a `boolean` for determining
whether to insert the noun with a new Bag<Integer> containing its
current synset, or whether to append the current synset to the existing
Bag<Integer> contained within the RedBlackBST.