Fix family-checking bug
Previously, two things happened:
(1) Family-checking did not occur during the BFS initialization phase.  This was an oversight,
since many of the words in the Bag of words are closely related, and are very likely to land close to each
other in the WordNet, and from there travel upwards to a very close ancestor (but not an ancestor with
a word from the other family).
(2) Family-checking did not occur after BFS initialization, when the search itself was occurring.
This meant that close ancestors were being returned without regard to whether they were from V or W.

Accordingly, assign families to the synsets from the different iterables (V and W), and use this
auxiliary data structure to check whether an ancestor has been found, since a word is not an ancestor
unless it has roots in a synset from V and in a synset from W.