Keep track of family visitation separately
Previously, the data structure used for visiting synsets was a single `distTo` array.

However, this does not allow for keeping track of the shortest path to a given synset from multiple paths,
as is the case whenever the digraph has cycles.

Accordingly, it is preferable to keep track of `distTo` in two separate arrays: one that tracks the distance
to that synset from one of the families, and another for the distance to that synset from the other family.

Whenever a family begins to loop, that family can cease to enqueue new synsets.

However, so long as that family has not explored a set of synsets, it can continue to enqueue new synsets, since it may
discover a distance-saving loop.

As mentioned in the checklist, we know that we can terminate the BFS once our depth exceeds the length of the best
shortest ancestral path found so far.  This is because sap is often the sum of two paths, and so its length can be
at most twice the current depth, but since there's always the possibility that one of the two paths finds a loop
directly to the start of the other family, it must keep exploring.